refactor(page): extract useHolePunchManager hook (phase 3)
Move the hole-punch subsystem — placement state/refs, click/hover/select,
drag + gizmo move/rotate handlers, the ~276-line apply, auto-depth, and
the delete-handler registration — out of page.tsx into
features/hole-punching/useHolePunchManager.ts (reusing the Phase-1
holePunchGeometry/holePunchPersistence helpers). State lives in the hook,
consumed in Home via destructure-same-names so all consumers
(HolePunchPanel/Gizmo/PreviewCylinder, SceneCanvas overlays) are unchanged.

Shared modifier-reset/blocker machinery and the hollowing state it
cross-reads stay in Home and are injected as deps (28 deps); 47 names
returned. Authored by a worktree subagent, integrated + re-verified here.

page.tsx: 20,195 -> 19,133 lines (-1,062). Verified via `tsc --noEmit`.
Smoke-test hole-punch placement/gizmo/apply before merge.
